+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+-   Added the `trigger:create:scheduler` and `trigger:delete:scheduler` commands to allow creating/deleting a Cloud Scheduler job to trigger the Cloud Run service. This supports both authenticated and unauthenticated invocations, and will create/setup service account as needed. `trigger:delete:scheduler` is invoked automatically with `crwt destroy`. This is the first of other potential trigger:[create/delete]:{method} triggers.
+-   Changed the way the `cloudrun` section within the `.crbt` configuration file is stored. Previously used a more complex approach based on earlier prototype approaches. The `parseConfig` library and references were updated to reflect.
+-   Documentation and formatting fixes.
+
 ## [0.1.1] - 2020-02-13
 
 ### Added